requirements.txt
numpy
pillow
concurrent-futures
memory-profiler
gc
PyYAML